% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jpaper-package.R
\docType{package}
\name{jpaper-package}
\alias{jpaper}
\alias{jpaper-package}
\title{jpaper: What the Package Does (One Line, Title Case)}
\description{
Personal package.
}
\author{
\strong{Maintainer}: Dongdong Kong \email{kongdd.sysu@gmail.com}

}
\keyword{internal}
