% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/t_test.R
\name{t_test}
\alias{t_test}
\alias{t_test_simple}
\alias{t_test_simple2}
\title{T test and TukeyHSD test}
\usage{
t_test(x, y)

t_test_simple(file, outfile = NULL)

t_test_simple2(file, outfile = "result.xlsx")
}
\arguments{
\item{x}{A numeric vector}

\item{y}{A numeric vector}

\item{file}{The file path of xlsx}

\item{outfile}{The file path of xlsx for OUTPUTS}
}
\description{
T test and TukeyHSD test
}
\examples{
set.seed(1)
x <- rnorm(100)
y <- rnorm(100)
t_test(x, y)

\dontrun{
t_test_simple("a.xlsx")
}
\dontrun{
t_test_simple2(file)
}
}
\seealso{
\code{\link[stats:t.test]{stats::t.test()}}, \code{\link[stats:TukeyHSD]{stats::TukeyHSD()}}
}
