% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/t_test.R
\name{t_test}
\alias{t_test}
\alias{t_test_simple}
\title{T test and TukeyHSD test}
\usage{
t_test(x, y)

t_test_simple(file, outfile = NULL)
}
\arguments{
\item{x}{A numeric vector}

\item{y}{A numeric vector}

\item{file}{The file path of xlsx}
}
\description{
T test and TukeyHSD test
}
\examples{
\dontrun{
t_test(x, y)
t_test_simple("a.xlsx")
}
}
\seealso{
\code{\link[stats:t.test]{stats::t.test()}}, \code{\link[stats:TukeyHSD]{stats::TukeyHSD()}}
}
